Android client can now properly register new user and login. No error handling yet.
Added CmdType class to both client and server so that web protocol can be clearly specified there.
Added Message.java class. The class is used to store output message
Reworked Connection class
Changes to Acticity classes so they are compatible with new Connection class
Changes to WebHandler.py so that it uses CmdType